requirements.txt
asgiref
black
cachetools
certifi
charset-normalizer
click
colorama
Django
django-gmailapi-backend
django-taggit
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
googleapis-common-protos
httplib2
idna
mypy-extensions
pathspec
platformdirs
protobuf
pyasn1
pyasn1-modules
pyparsing
requests
rsa
six
sqlparse
tomli
tzdata
uritemplate
urllib3
whitenoise
